# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "ClockWise is a workforce management and scheduling platform for organizations.",
        "title": "ClockWise API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "ClockWise API Support",
            "email": "ziadeliwa@aucegypt.edu"
        },
        "license": {
            "name": "MIT License",
            "url": "https://opensource.org/license/mit"
        },
        "version": "1.0.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
...
        "/register": {
            "post": {
                "description": "Creates a new organization along with an admin user account",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Organizations"
                ],
                "summary": "Register a new organization",
                "parameters": [
                    {
                        "description": "Organization registration details",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.RegisterOrgRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Organization registered successfully",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Invalid request body",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Registration failed",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/{org}/staffing": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns staffing overview including total employees and breakdown by role",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Staffing"
                ],
                "summary": "Get staffing summary",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Organization ID",
                        "name": "org",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Staffing summary",
                        "schema": {
                            "$ref": "#/definitions/api.StaffingSummary"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Failed to retrieve employees",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Delegates/creates a new user in the organization. Only admins and managers can perform this action.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Staffing"
                ],
                "summary": "Create a new employee",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Organization ID",
                        "name": "org",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "New employee details",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.DelegateUserRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "User delegated successfully",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Invalid request body",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "403": {
                        "description": "Staff cannot delegate users",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "api.DelegateUserRequest": {
            "type": "object",
            "required": [
                "email",
                "full_name",
                "role"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "full_name": {
                    "type": "string"
                },
                "role": {
                    "type": "string",
                    "enum": [
                        "manager",
                        "staff"
                    ]
                }
            }
        },
...
        "api.RegisterOrgRequest": {
            "type": "object",
            "required": [
                "admin_email",
                "admin_full_name",
                "admin_password",
                "org_name"
            ],
            "properties": {
                "admin_email": {
                    "type": "string"
                },
                "admin_full_name": {
                    "type": "string"
                },
                "admin_password": {
                    "type": "string",
                    "minLength": 6
                },
                "org_address": {
                    "type": "string"
                },
                "org_name": {
                    "type": "string"
                }
            }
        },
...
        "api.StaffingSummary": {
            "type": "object",
            "properties": {
                "by_role": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "integer"
                    }
                },
                "employees": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/database.User"
                    }
                },
                "total_employees": {
                    "type": "integer"
                }
            }
        },
        "database.User": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "full_name": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "organization_id": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "user_role": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
...
}